functions/bunq_money_flow/requirements.txt
aenum==2.2.6 ; python_full_version >= "3.11.0" and python_full_version < "3.12.0"
blinker==1.7.0 ; python_full_version >= "3.11.0" and python_full_version < "3.12.0"
bunq-sdk==1.14.18 ; python_full_version >= "3.11.0" and python_full_version < "3.12.0"
cachecontrol==0.13.1 ; python_full_version >= "3.11.0" and python_full_version < "3.12.0"
cachetools==5.3.2 ; python_full_version >= "3.11.0" and python_full_version < "3.12.0"
certifi==2023.11.17 ; python_full_version >= "3.11.0" and python_full_version < "3.12.0"
cffi==1.16.0 ; python_full_version >= "3.11.0" and python_full_version < "3.12.0"
chardet==3.0.4 ; python_full_version >= "3.11.0" and python_full_version < "3.12.0"
charset-normalizer==3.3.2 ; python_full_version >= "3.11.0" and python_full_version < "3.12.0"
click==8.1.7 ; python_full_version >= "3.11.0" and python_full_version < "3.12.0"
cloudevents==1.9.0 ; python_full_version >= "3.11.0" and python_full_version < "3.12.0"
colorama==0.4.6 ; python_full_version >= "3.11.0" and python_full_version < "3.12.0" and platform_system == "Windows"
cryptography==41.0.7 ; python_full_version >= "3.11.0" and python_full_version < "3.12.0"
deprecation==2.1.0 ; python_full_version >= "3.11.0" and python_full_version < "3.12.0"
firebase-admin==6.3.0 ; python_full_version >= "3.11.0" and python_full_version < "3.12.0"
firebase-functions==0.2.0 ; python_full_version >= "3.11.0" and python_full_version < "3.12.0"
flask-cors==4.0.0 ; python_full_version >= "3.11.0" and python_full_version < "3.12.0"
flask==3.0.0 ; python_full_version >= "3.11.0" and python_full_version < "3.12.0"
functions-framework==3.5.0 ; python_full_version >= "3.11.0" and python_full_version < "3.12.0"
google-api-core==2.15.0 ; python_full_version >= "3.11.0" and python_full_version < "3.12.0"
google-api-core[grpc]==2.15.0 ; python_full_version >= "3.11.0" and python_full_version < "3.12.0"
google-api-python-client==2.111.0 ; python_full_version >= "3.11.0" and python_full_version < "3.12.0"
google-auth-httplib2==0.2.0 ; python_full_version >= "3.11.0" and python_full_version < "3.12.0"
google-auth==2.26.1 ; python_full_version >= "3.11.0" and python_full_version < "3.12.0"
google-cloud-core==2.4.1 ; python_full_version >= "3.11.0" and python_full_version < "3.12.0"
google-cloud-firestore==2.14.0 ; python_full_version >= "3.11.0" and python_full_version < "3.12.0"
google-cloud-secret-manager==2.17.0 ; python_full_version >= "3.11.0" and python_full_version < "3.12.0"
google-cloud-storage==2.14.0 ; python_full_version >= "3.11.0" and python_full_version < "3.12.0"
google-crc32c==1.5.0 ; python_full_version >= "3.11.0" and python_full_version < "3.12.0"
google-events==0.11.0 ; python_full_version >= "3.11.0" and python_full_version < "3.12.0"
google-resumable-media==2.7.0 ; python_full_version >= "3.11.0" and python_full_version < "3.12.0"
googleapis-common-protos==1.62.0 ; python_version >= "3.11" and python_full_version < "3.12.0"
googleapis-common-protos[grpc]==1.62.0 ; python_full_version >= "3.11.0" and python_full_version < "3.12.0"
grpc-google-iam-v1==0.13.0 ; python_full_version >= "3.11.0" and python_full_version < "3.12.0"
grpcio-status==1.60.0 ; python_version >= "3.11" and python_full_version < "3.12.0"
grpcio==1.60.0 ; python_version >= "3.11" and python_full_version < "3.12.0"
gunicorn==21.2.0 ; python_full_version >= "3.11.0" and python_full_version < "3.12.0" and platform_system != "Windows"
httplib2==0.22.0 ; python_full_version >= "3.11.0" and python_full_version < "3.12.0"
idna==3.6 ; python_full_version >= "3.11.0" and python_full_version < "3.12.0"
itsdangerous==2.1.2 ; python_full_version >= "3.11.0" and python_full_version < "3.12.0"
jinja2==3.1.2 ; python_full_version >= "3.11.0" and python_full_version < "3.12.0"
markupsafe==2.1.3 ; python_full_version >= "3.11.0" and python_full_version < "3.12.0"
msgpack==1.0.7 ; python_full_version >= "3.11.0" and python_full_version < "3.12.0"
packaging==23.2 ; python_full_version >= "3.11.0" and python_full_version < "3.12.0"
proto-plus==1.23.0 ; python_version >= "3.11" and python_full_version < "3.12.0"
protobuf==4.25.1 ; python_version >= "3.11" and python_full_version < "3.12.0"
pyasn1-modules==0.3.0 ; python_full_version >= "3.11.0" and python_full_version < "3.12.0"
pyasn1==0.5.1 ; python_full_version >= "3.11.0" and python_full_version < "3.12.0"
pycparser==2.21 ; python_full_version >= "3.11.0" and python_full_version < "3.12.0"
pycryptodomex==3.19.1 ; python_full_version >= "3.11.0" and python_full_version < "3.12.0"
pyjwt[crypto]==2.8.0 ; python_full_version >= "3.11.0" and python_full_version < "3.12.0"
pyparsing==3.1.1 ; python_full_version >= "3.11.0" and python_full_version < "3.12.0"
python-dotenv==1.0.0 ; python_full_version >= "3.11.0" and python_full_version < "3.12.0"
pyyaml==6.0.1 ; python_full_version >= "3.11.0" and python_full_version < "3.12.0"
requests==2.31.0 ; python_full_version >= "3.11.0" and python_full_version < "3.12.0"
rsa==4.9 ; python_full_version >= "3.11.0" and python_full_version < "3.12.0"
simplejson==3.19.2 ; python_full_version >= "3.11.0" and python_full_version < "3.12.0"
typing-extensions==4.9.0 ; python_full_version >= "3.11.0" and python_full_version < "3.12.0"
uritemplate==4.1.1 ; python_full_version >= "3.11.0" and python_full_version < "3.12.0"
urllib3==1.26.18 ; python_full_version >= "3.11.0" and python_full_version < "3.12.0"
watchdog==3.0.0 ; python_full_version >= "3.11.0" and python_full_version < "3.12.0"
werkzeug==3.0.1 ; python_full_version >= "3.11.0" and python_full_version < "3.12.0"
